make every version imply the one before it
This allows the library to be compiled with any combination of versions
of Bitcoin Core -- it will just use the highest one.

This won't guarantee compatibility because Bitcoin Core sometimes makes
backward-compatibility breaking changes ... but it is much better than
the current situation where if you specify multiple feature flags, the
library won't compile at all.