Use range dependencies
The last couple of releases of `bitcoincore-rpc` and `bitcoin_hashes`
have not required code changes here. This means we can use range
dependencies. The benefit is that it makes the upgrade of other crates
that use `bitcoind` easier.

Tested by pinning `hashes` with:

- cargo update -p bitcoin_hashes@0.13.0 --precise 0.13.0
- cargo update -p bitcoin_hashes@0.14.0 --precise 0.14.0

And then pinning `bitcoincore-rpc` with all of the dependencies in the
range and running `cargo check --all --all-features --all-targets`

E.g (using `cc` as alias to the cargo check command above)

- cargo update -p bitcoincore-rpc --precise 0.17.0 && cc
- cargo update -p bitcoincore-rpc --precise 0.18.0 && cc
- cargo update -p bitcoincore-rpc --precise 0.19.0 && cc

Bump the crate patch version so we can release this change in a point
release. Since this is a pre-1.0 crate I believe this is ok to do.